requirements.txt
fastapi
uvicorn
gradio
pydantic
python-dotenv
pymongo
resend
requests
cloudinary
pillow